gst123: fix crash in window resizing code, fix player show/hide for gstreamer-1.0 builds
Before this fix, the caps_set_cb function would perform window resizing,
although it is only safe to do this from the  main. Result: it sometimes
crashed. This was fixed.

On gstreamer-1.0, the window would not show()/hide() depending on whether
an audio file was played, or a video file (or audio + visualization). This
should now work for gstreamer-0.10 and gstreamer-1.0.