add support to change playback rate and direction
This is mostly the keybindings from mplayer, except it doesn't have
play in reverse.

Reverse isn't a freebee with gst_element_seek, that is you can't just
change the sign of the rage.  If stop/start are also flipped and
without using set the stream can terminate based on the previous set
value.